-Correction d'un bug lors de la récupération des classifications des fichiers laz, les 8 bits de l'octet étaient récupérés au lieu des 5 bits de poids le plus faible. Conséquence: classification incorrecte -Déplacement des classes liées à la lecture des fichiers las et laz dans le package fr.amap.lidar.als -Modification du projet Qt LasZipLibrary par rapport au déplacement des classes dans un autre package -Création d'une fonction permettant de récupérer une sous-partie d'un MNT -Lors du chargement de plusieurs fichiers las, ajout de la possibilité de générer un raster de mnt par fichier las à partir d'un raster global

+++ trunk/VoxeLidarGUIFX/src/main/java/fr/ird/voxelidar/util/ByteConverter.java
@@ -19,8 +19,8 @@ public class ByteConverter {
         return (int)(unsignedShort & 0xffff);
     }
     
-    public static short unsignedByteToShort(byte unsignedByte){
-        return (short)(unsignedByte & 0xff);
+    public static int unsignedByteToShort(byte unsignedByte){
+        return (int)(unsignedByte & 0xff);
     }
     
     public static long unsignedIntegerToLong(int unsignedInteger){
